Refactor arquitectura, paquetes, relaciones, quitamos security
