+++ src/main/java/com/alkemy/wallet/controller/exception/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.alkemy.wallet.model.exception;
+package com.alkemy.wallet.controller.exception;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
