+++ src/main/java/com/alkemy/wallet/model/dto/response/FixedTermDepositResponseDto.java
@@ -1,6 +1,5 @@
-package com.alkemy.wallet.model.response;
+package com.alkemy.wallet.model.dto.response;
 
-import com.alkemy.wallet.model.TransactionTypeEnum;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -10,17 +9,17 @@ import java.time.LocalDateTime;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class TransactionResponseDto {
+public class FixedTermDepositResponseDto {
     private Long transactionsId;
     private Double amount;
-    private TransactionTypeEnum type;
-    private String description;
     private Long userId;
     private Long accountId;
-    private LocalDateTime transactionDate;
+    private Double interest;
+    private LocalDateTime creationDate;
+    private LocalDateTime closingDate;
 
-    private UserResponseDto userResponseDto;
     private AccountResponseDto accountResponseDto;
+    private UserResponseDto userResponseDto;
 
     public Long getTransactionsId() {
         return transactionsId;
@@ -36,22 +35,6 @@ public class TransactionResponseDto {
 
     public void setAmount(Double amount) {
         this.amount = amount;
-    }
-
-    public TransactionTypeEnum getType() {
-        return type;
-    }
-
-    public void setType(TransactionTypeEnum type) {
-        this.type = type;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
     }
 
     public Long getUserId() {
@@ -70,20 +53,28 @@ public class TransactionResponseDto {
         this.accountId = accountId;
     }
 
-    public LocalDateTime getTransactionDate() {
-        return transactionDate;
+    public Double getInterest() {
+        return interest;
     }
 
-    public void setTransactionDate(LocalDateTime transactionDate) {
-        this.transactionDate = transactionDate;
+    public void setInterest(Double interest) {
+        this.interest = interest;
     }
 
-    public UserResponseDto getUserResponseDto() {
-        return userResponseDto;
+    public LocalDateTime getCreationDate() {
+        return creationDate;
     }
 
-    public void setUserResponseDto(UserResponseDto userResponseDto) {
-        this.userResponseDto = userResponseDto;
+    public void setCreationDate(LocalDateTime creationDate) {
+        this.creationDate = creationDate;
+    }
+
+    public LocalDateTime getClosingDate() {
+        return closingDate;
+    }
+
+    public void setClosingDate(LocalDateTime closingDate) {
+        this.closingDate = closingDate;
     }
 
     public AccountResponseDto getAccountResponseDto() {
@@ -92,5 +83,13 @@ public class TransactionResponseDto {
 
     public void setAccountResponseDto(AccountResponseDto accountResponseDto) {
         this.accountResponseDto = accountResponseDto;
+    }
+
+    public UserResponseDto getUserResponseDto() {
+        return userResponseDto;
+    }
+
+    public void setUserResponseDto(UserResponseDto userResponseDto) {
+        this.userResponseDto = userResponseDto;
     }
 }
