+++ src/main/java/com/alkemy/wallet/model/mapper/response/complemento/IAccountResponse2Mapper.java
@@ -1,7 +1,7 @@
 package com.alkemy.wallet.model.mapper.response.complemento;
 
 import com.alkemy.wallet.model.entity.Account;
-import com.alkemy.wallet.model.response.AccountResponseDto;
+import com.alkemy.wallet.model.dto.response.AccountResponseDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
