+++ src/main/java/com/alkemy/wallet/model/mapper/response/role/IRoleResponseMapper.java
@@ -1,7 +1,7 @@
 package com.alkemy.wallet.model.mapper.response.role;
 
 import com.alkemy.wallet.model.entity.Role;
-import com.alkemy.wallet.model.response.RoleResponseDto;
+import com.alkemy.wallet.model.dto.response.RoleResponseDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
